Add files via upload
This tool was created to allow the KF-RNAseq-Workflow to be run via the GUI in CAVATICA. To create a batch analysis, batched inputs must go through one input port. This tool allows both R1 and R2 (if paired-end analysis) to be input to the same port (to allow batching) and splits them into R1 and R2 based on their metadata. If only a single file is needed then it is passed through and the remainder of the workflow is unaltered.